in progress: globalize player / adventurer counting
