+++ services/app/src/reducers/StateTypes.tsx
@@ -84,7 +84,7 @@ export interface SettingsType {
   experimental: boolean;
   fontSize: FontSizeType;
   multitouch: boolean;
-  numPlayers: number;
+  numLocalPlayers: number;
   showHelp: boolean;
   simulator: boolean;
   timerSeconds: number;
